fix: add info tooltip
